py/requirements.txt
cobs==1.2.1
contourpy==1.1.1
cycler==0.12.1
fonttools==4.46.0
importlib-resources==6.1.1
kiwisolver==1.4.5
matplotlib==3.7.4
numpy==1.24.4
packaging==23.2
Pillow==10.1.0
pygame==2.5.2
pyparsing==3.1.1
PyQt5==5.15.10
PyQt5-Qt5==5.15.11
PyQt5-sip==12.13.0
pyqtgraph==0.13.3
pyserial==3.5
python-dateutil==2.8.2
robust-serial==0.2
six==1.16.0
zipp==3.17.0
